Move ConfigurationForm to its own component
